Move ResNet to core
